requirements.txt
spacy~=3.5.3
requests~=2.31.0
python-dotenv~=1.0.0